cli and gui:  When applying from a file that was created for a different product, make sure to reset the hard current limit to their defaults (and also set calibration to 0).
Motivation: We want to distribute file for our linear actuators that
work on multiple Jrks.  At best, it is tricky to figure out what value to
write for the encoded_hard_current_limit in those files that will
be valid for all the models we care about.

New Jrk API functions:

  jrk_settings_fix_and_change_product
  jrk_settings_set_firmware_version
  jrk_settings_get_firmware_version

The firmware version will eventually be needed in the object so we
can use it in jrk_settings_fix to disable features that are not available
yet.  Also, it should be an argument to jrk_settings_fix_and_change_product
for that same reason.  Since we are adding jrk_settings_fix_and_change_product
now, we might as well go ahead and fully implement tracking of the
firmware version in the object too, even though it isn't used.

Code that was audited/adapted to make sure it uses these new API
functions properly:

- lib:
  - defaults -> object jrk_settings_fill_with_defaults  : don't clear firmware version,
                                                          always set hard_overcurrent_threshold
  - EEPROM -> object   jrk_get_eeprom_settings          : set firmware version
  - RAM -> object      jrk_get_ram_settings             : set firmware version
  - string -> object   jrk_settings_read_from_string    : no change
  - object -> EEPROM   jrk_set_eeprom_settings          : use jrk_settings_fix_and_change_product
  - object -> RAM      jrk_set_eeprom_settings          : use jrk_settings_fix_and_change_product
  - object -> string   jrk_settings_to_string           : no change; don't print firmware version
- cli:
  - EEPROM -> string   get_eeprom_settings              : no change
  - RAM -> string      get_ram_settings                 : no change
  - string -> EEPROM   set_eeprom_settings              : use jrk_settings_fix_and_change_product
  - string -> RAM      set_ram_settings                 : use jrk_settings_fix_and_change_product
  - string -> string   fix_settings                     : no change
- gui:
  - EEPROM -> object   reload_settings, connect_device  : no change
  - string -> object   open_settings_from_file          : use jrk_settings_fix_and_change_product
  - object -> EEPROM   apply_settings                   : assert the product/version matches
  - object -> string   save_settings_to_file            : no change

Also, this commit fixes several places where the product was used or stored
as a uint8_t instead of uint32_t.

Also, updated the version number to 1.3.1 and edited the README.

+++ lib/jrk_device.c
@@ -9,7 +9,7 @@ struct jrk_device
   char * serial_number;
   char * os_id;
   uint16_t firmware_version;
-  uint8_t product;
+  uint32_t product;
 };
 
 jrk_error * jrk_list_connected_devices(
